LayoutTests/fast/notifications/notifications-request-permission-optional.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script type="text/javascript">
shouldBeDefined("window.webkitNotifications");
shouldNotThrow("window.webkitNotifications.requestPermission()");
</script>
<script src="../../resources/js-test-post.js"></script>
</head>
</html>
